backend/pyproject.toml
[project]
name = "serpentstack-backend"
version = "0.1.0"
description = "SerpentStack backend API - production-ready FastAPI template for AI-agent-assisted development"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "slowapi>=0.1.9",
    "structlog>=24.4.0",
    "PyJWT[crypto]>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.0.0",
    "httpx>=0.28.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "arq>=0.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "testcontainers[postgres]>=4.0.0",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "UP",    # pyupgrade
    "RUF",   # ruff-specific rules
]
ignore = [
    "B008",  # function call in default arg — FastAPI Depends()/Security() require this
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
